feat: 添加账单功能完成

+++ app/src/main/java/com/ark/truckbill/data/Bill.kt
@@ -5,8 +5,8 @@ import java.time.LocalDate
 
 data class Bill(
     val name: String,
-    val weight: Float,
-    val price: Float,
+    val weight: Int,
+    val price: Int,
     val start: String,
     val end: String,
     val startDate: LocalDate,
